requirements.txt
pyyaml>=6.0.1
typing-extensions>=4.8.0
openai>=1.3.5
llama-cpp-python>=0.2.6
aiohttp>=3.8.6
PyGithub==2.1.1
python-dotenv>=1.0.0
